requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
websockets>=12.0
aiohttp>=3.9.1
curl_cffi>=0.7.0
sqlalchemy>=2.0.30
aiosqlite>=0.20.0
pyjwt>=2.8.0
slowapi>=0.1.9
asyncpg>=0.29.0
gunicorn>=21.0.0
redis>=5.0.0
arq>=0.25.0
